+++ libraries/AP_AnalogSource/AP_AnalogSource_ADC.h
@@ -10,7 +10,7 @@ class AP_AnalogSource_ADC : public AP_AnalogSource
     public:
     AP_AnalogSource_ADC( AP_ADC * adc, int ch, float prescale = 1.0 ) :
         _adc(adc), _ch(ch), _prescale(prescale) {}
-    int read(void);
+    float read(void);
 
     private:
     AP_ADC * _adc;
